feat: add WhiteHeaderContainer widget and Union image
